complete conversion from float to double, tested accuracy to "11 digits of accuracy in 4425 ticks"

+++ BingoBlockParty/BingoBlockParty.Client/BallGame/Pieces/ClientCannon.cs
@@ -103,7 +103,7 @@ namespace BingoBlockParty.Client.BallGame.Pieces
                     context.Translate(cannonLocation.X - cannonImage.Width / 2, cannonLocation.Y);
 
 
-                    context.DrawImage(cannonImage, 0, 0, (float)(this.GameBoard.GameModel.CannonAngle * Math.PI / 180), cannonImage.Width / 2, 0);
+                    context.DrawImage(cannonImage, 0, 0,  (this.GameBoard.GameModel.CannonAngle * Math.PI / 180), cannonImage.Width / 2, 0);
               
                     context.Restore();
